+++ src/LDL/Framework/Base/Contracts/HasKeyInterface.php
@@ -1,0 +1,11 @@
+<?php declare(strict_types=1);
+
+namespace LDL\Framework\Base\Contracts;
+
+interface HasKeyInterface
+{
+    /**
+     * @return string
+     */
+    public function getObjectKey() : string;
+}